# cmd/monitor-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a monitor gateway",
        "title": "monitor-gateway",
        "termsOfService": "http://www.swagger.io/support",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/monitor-gateway/api/v1/demo/hello-world": {
            "post": {
                "description": "查询用户第一条数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FetchFirst"
                ],
                "summary": "FetchFirst",
                "operationId": "FetchFirst",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_router_demo.UserBody"
                        }
                    },
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "userID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_xiaoxlm_monitor-gateway_api_response.UserRESP"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorRESP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorRESP"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api_router_demo.UserBody": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "salary": {
                    "type": "number"
                }
            }
        },
        "github_com_xiaoxlm_monitor-gateway_api_response.UserRESP": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "httputil.ErrorRESP": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                }
            }
        }
    }
}